Update button background colors
We don't want people accidentally hitting the "revert to default" button after just adjusting several colors to their liking only to have their work destroyed.  Making the "save and update" button green and the "revert" button red can majorly help people to avoid this button and not confuse it with the other.