Add intentional delay to color randomization
When the colors on the snail are randomized, the MySnail data in the Chrome User's Sync is updated across all devices.  This can only be done a certain number of times per minute.  To avoid reaching or exceeding this limit, a delay of two seconds (or more, for AI training and snail judging, etc.) runs each time the "Randomize" button is clicked.