FIX BUG: Save settings when slider at endpoint
Previously, when the slider for setting the snail's size or speed was dragged to other endpoint but the mouse button was released while the mouse was not on the slider, the new setting would not save.  Now, when previewing a maximum/minimum size or speed, as soon as the slider hits the end the setting will be saved to Chrome sync.